style & styles Div
