--- packages/ui/src/HTML/React/style/Div.tsx
@@ -1,11 +1,0 @@
-import styled, { css } from "styled-components";
-
-type DivStyled = {
-  empty?: boolean;
-};
-
-const emptyDiv = css<DivStyled>``;
-
-export const Div = styled.div<DivStyled>`
-  ${props => props.empty && emptyDiv}
-`;
