+++ packages/ui/src/HTML/React/style/HTMLElements/index.ts
@@ -3,3 +3,4 @@ export * from "./P";
 export * from "./Input";
 export * from "./Img";
 export * from "./H1";
+export * from "./Div";
