+++ packages/ui/src/HTML/index.ts
@@ -1,5 +1,5 @@
 import * as React from "./React";
-export { Input, P, Path, Img, H1 } from "./React";
+export { Input, P, Path, Img, H1, Div } from "./React";
 import * as Next from "./Next/index";
 import * as HtmlClasses from "./class";
