+++ packages/ui/src/pages/styles/html-elements/div/index.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import { Div } from "../../../../HTML";
+
+export default function index() {
+  return <Div>Sono un Div</Div>;
+}
